% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_aves}
\alias{atlas_aves}
\title{Data from Atlas das Aves Nidificantes de Portugal Continental}
\format{A data frame with 8 columns x 71111 rows:
\describe{
  \item{id_ap_bio3}{unique internal id}
  \item{especie}{Scientific name of the species}
  \item{cod_nidificacao}{Atlas reproduction status code}
  \item{utm}{UTM Grid Code of occurence}
}}
\usage{
atlas_aves
}
\description{
This is data from 2005 Atlas (Equipa Atlas).
}
\keyword{datasets}
