% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulFlora.R
\name{tabulFlora}
\alias{tabulFlora}
\title{Build report-ready table with all known bird occurences for a region of interest.}
\usage{
tabulFlora(
  utm_ae = utm_ae,
  utm_q = utm_contig,
  fielddata = NULL,
  biblio = "all"
)
}
\arguments{
\item{utm_ae}{character vector of utm grid cells touching roi.}

\item{utm_q}{character vector of utm grid cells around roi}

\item{fielddata}{a dataframe with bid data obtained for the roi. Tipically a table with species
utm reference.}

\item{biblio}{character vector. Databases to get data from.  One of c('all', 'floraon', 'pterid').}

\item{ae}{any spatial (sp, spatialdataframe) or coordinates vector indicating a region
of interest.}
}
\value{
an Excel file with a complete and pre-formatted table,
ready for use on reports.
}
\description{
The \code{tabulAve} function will create a reference table (Excel file)
considering all bird species occuring in a roi, joining all relevant information
(taxonomic, conservation, occurence status and legal framework).
It will also summarize roi status (ratios for protected and threatened species).
}
\details{
ensure that field datasheet follow the standardized formatting.
}
\examples{
   # read roi from shapefile
   ae = sf::read_sf(here::here('sig', 'ae_buffer250m.shp'),
                    stringsAsFactors = F) \%>\%
    st_set_crs(3763)
   # Cast multipolygon geometry to single parts
   aeu = st_cast(ae, "POLYGON")
   # Add an UID to each
   aeu$id = c(1:2)
   # get UTM codes for your roi
   utm_all = utm_id(grid = utm10k, roi = aeu \%>\% filter(id == 2), buff = NULL, contiguity = 'queen')
   # tabulate species occurence and status in the area
   tflora = tabulEIA::tabulFlora(fielddata = NULL, utm_ae = utm_ae, utm_q = utm_contig,
                               biblio = 'all')
   # export to csv
}
\author{
Paulo E. Cardoso
}
