% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulHerp.R
\name{tabulHerp}
\alias{tabulHerp}
\title{Build report-ready table with all known amphibians and reptiles occurrences for a region of interest.}
\usage{
tabulHerp(
  utm_ae = utm_ae,
  utm_q = utm_contig,
  fielddata = NULL,
  atlas = atlas_herp,
  dhab = dh13_18,
  inat = NULL
)
}
\arguments{
\item{utm_ae}{character. vector of utm grid cells touching roi.}

\item{utm_q}{character. vector of utm grid cells around roi}

\item{fielddata}{dataframe. Table with bid data obtained for the roi. Typically a table with species
utm reference.}

\item{atlas}{character. ICNF Atlas data with references 17384 entries to the UTM 10x10 grid cell.}

\item{dhab}{character. Use Habitat Directive data from 2013-2018 report published by ICNF.}

\item{inat}{logical. TRUE to use Inaturalist Dataset for 2018 and 2019 with thousands of entries available
on iNaturalist database from 2018 and 2019.}

\item{ae}{sf object. any spatial (sp, spatialdataframe) or coordinates vector indicating a region
of interest.}
}
\value{
an Excel file with a complete and pre-formatted table,
(almost) ready for use on reports.
}
\description{
The \code{tabulMam} function will create a reference table (Excel file)
considering all amphibians and reptiles species occurring in a region, joining all relevant information
(taxonomic, conservation, occurrence status and legal framework).
It will also (in the future) summarize region status (ratios for protected and threatened species).
}
\details{
ensure that field dataset follow the standardized formatting.
}
\examples{
\dontrun{
# read roi from shapefile
ae <- sf::read_sf(here::here("sig", "ae_buffer250m.shp"),
  stringsAsFactors = F
) \%>\%
  st_set_crs(3763)
# Cast multipolygon geometry to single parts
aeu <- st_cast(ae, "POLYGON")
# Add an UID to each
aeu$id <- c(1:2)
# get UTM codes for your roi
utm_all <- utm_id(grid = utm10k, roi = aeu \%>\% filter(id == 2), buff = NULL, contiguity = "queen")
# tabulate species occurence and status in the area
tmam <- tabulEIA::tabulMam(utm_ae = utm_ae, utm_q = utm_contig, fielddata = NULL, atlas = "all", inat = FALSE)
# export to csv
write_excel_csv(tmam, path = here::here("output", "tab_mammals.csv"))
}
}
\author{
Paulo E. Cardoso
}
