% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utmID.R
\name{utm_id}
\alias{utm_id}
\title{Identify the UTM grid squares touching and around a region of interest (roi).}
\usage{
utm_id(grid = utm10, roi = ae, buf = NULL, contiguity = "queen")
}
\arguments{
\item{grid}{character indicating the UTM grid. Ensure you provide an 'UTM' column as
grid code.}

\item{roi}{an spatial object sf, st, spatialpolygonsdataframe identifying
the roi.' It can be of any valid vector type (point, line, polygon or collection).}

\item{buf}{numeric indication whether the roi should be buffered to a given distance (meter).
Default is NULL}

\item{contiguity}{character identifying the contiguity search method "queen" or "rook".}
}
\value{
A vector with UTM codes for study site (ae) and contiguous grids.
}
\description{
The \code{utm_id} function will identify the UTM grid codes
that touches and those in the vicinity (queen or rook) of roi.
}
\details{
Make sure you have a topologically valid  vector. Using external files (like a shapefile)
will require the .prj otherwise identify it with the corresponding numeric epsg code.
Currently only rook and queen implemented (not bishop)
}
\examples{
\dontrun{
   require(tabulEIA)
   require(sf)
   require(tidyverse)
   # get UTM codes for your roi
   utm_all = utm_id(grid = utm10k, roi = ae, buf = NULL, contiguity = 'queen')
   utm_ae <- utm_all$ae
   utm_contig <- utm_all$contig
   }
check projection supplied
buffer roi
Intersects
}
\author{
Paulo E. Cardoso
}
