% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utmID.R
\name{utm_id}
\alias{utm_id}
\title{Identify the UTM grid squares touching and around a region of interest (roi).}
\usage{
utm_id(grid = utm10, roi = ae, buff = NULL, contiguity = "queen")
}
\arguments{
\item{grid}{character indicating the UTM grid. Ensure you provide an 'UTM' column as
grid code.}

\item{roi}{an spatial object sf, st, spatialpolygonsdataframe identifying
the roi.
It can be of any valid vector type (point, line, polygon or collection).}

\item{contiguity}{character identifying the contiguity search method "queen" or "rook".
\code{\link[https://i.stack.imgur.com/CWIHi.jpg]{contiguity}}}

\item{buf}{numeric indication wether the roi should be buffered to a given distance (meter).
Default is NULL}
}
\value{
A vector with UTM codes.
}
\description{
The \code{utm_id} function will identify the UTM grid codes
that touches and those in the vicinity (queen or rook) of roi.
}
\details{
Make sure you have a topologically valid  vector. Using external files (like a shapefile)
will require the .prj otherwise identify it with the corresponding numeric epsg code.
}
\examples{
   # do not run
   # Read coodinates as organized for analysis of movements
   ae = "~path/to/roi/roi.shp"
   # identify the UTM touching and those from vicinity using 'queen' criteria
   utm_ae = utm_id(grid = utm, roi = ae, buff = 500, contiguity = 'queen')
}
\author{
Paulo E. Cardoso
}
